requirements.txt
Django==3.2.5
djangorestframework==3.12.4
psycopg2==2.9.5
python-dotenv==1.0.1
djangorestframework-simplejwt==4.7.2